+++ packages/core-js/internals/normalize-string-argument.js
@@ -1,0 +1,5 @@
+var toString = require('../internals/to-string');
+
+module.exports = function (argument, $default) {
+  return argument === undefined ? arguments.length < 2 ? '' : $default : toString(argument);
+};