+++ tests/tests/es.aggregate-error.js
@@ -8,6 +8,8 @@ QUnit.test('AggregateError', assert => {
   assert.ok(AggregateError([1]) instanceof AggregateError);
   assert.ok(AggregateError([1]) instanceof Error);
   assert.same(AggregateError([1], 'foo').message, 'foo');
+  assert.same(AggregateError([1], 123).message, '123');
+  assert.same(AggregateError([1]).message, '');
   assert.deepEqual(AggregateError([1, 2, 3]).errors, [1, 2, 3]);
   assert.throws(() => AggregateError([1], Symbol()), 'throws on symbol as a message');
 });